refactor(artifact-provenance): extract staging recovery
